🎉 Successful Scorecard Import & Update !!!
Updates Include:
---

- [X] Update Incorrect Cities and Counties
- [X] Add new `documents.scorecard` config option for CSV import
- [X] Add new `/update/scorecard` endpoint
- [X] Add new `csv-parse` npm package for parsing external CSV file
- [X] Update model definitions to fix import issues
- [X] Update sequelize dialect options for BIGINT support
- [X] Adds new util methods to standardize parsing CSV columns

+++ app/models/scorecard_police_funding.js
@@ -20,7 +20,7 @@ module.exports = (sequelize, DataTypes) => {
       comment: 'Foreign Key to scorecard_agency'
     },
     total_budget: {
-      type: DataTypes.INTEGER(10).UNSIGNED
+      type: DataTypes.BIGINT(15).UNSIGNED
     },
     police_budget: {
       type: DataTypes.INTEGER(10).UNSIGNED
